requirements.txt
websockets~=15.0
ultralytics~=8.3
torch~=2.7
numpy~=2.0.1
pillow~=11.2.1